refactor: client, server 분리
